+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning 2.0.0](https://semver.org/spec/
 ### Added
 
 - Advanced FTS Search, Gravity Sorting, and Pagination
+- DELETE post endpoint
 
 ## [2.0.0] - 2026-06-29